backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
firebase-admin==6.2.0
gspread==5.10.0
oauth2client==4.1.3
groq==0.4.1
scapy==2.5.0
pandas==2.0.3
joblib==1.3.2
APScheduler==3.10.4
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.5
reportlab==4.0.4
PyPDF2==3.0.1
fpdf==1.7.2
